rename storage to old_storage
