+++ src/agdb/collections/dictionary/dictionary_data_storage.rs
@@ -1,9 +1,9 @@
 use crate::collections::storage_multi_map::StorageMultiMap;
 use crate::collections::storage_vec::StorageVec;
 use crate::db::db_error::DbError;
-use crate::storage::storage_file::StorageFile;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_file::StorageFile;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
