+++ src/agdb/collections/map_common/map_data_storage.rs
@@ -2,8 +2,8 @@ use super::map_data::MapData;
 use super::map_value::MapValue;
 use super::map_value_state::MapValueState;
 use crate::db::db_error::DbError;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
