+++ src/agdb/collections/storage_map.rs
@@ -1,7 +1,7 @@
 use crate::db::db_error::DbError;
-use crate::storage::storage_file::StorageFile;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_file::StorageFile;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
