+++ src/agdb/collections/storage_multi_map.rs
@@ -4,9 +4,9 @@ use super::map_common::MapCommon;
 use super::multi_map::multi_map_impl::MultiMapImpl;
 use super::multi_map::MultiMap;
 use crate::db::db_error::DbError;
-use crate::storage::storage_file::StorageFile;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_file::StorageFile;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
