+++ src/agdb/collections/storage_vec.rs
@@ -1,8 +1,8 @@
 use super::vec::storage_vec_iterator::StorageVecIterator;
 use crate::db::db_error::DbError;
-use crate::storage::storage_file::StorageFile;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_file::StorageFile;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use crate::utilities::old_serialize::OldSerialize;
 use std::cell::RefCell;
 use std::cell::RefMut;
