+++ src/agdb/graph/storage_graph.rs
@@ -1,9 +1,9 @@
 use super::graph_data_storage::GraphDataStorage;
 use super::graph_impl::GraphImpl;
 use crate::db::db_error::DbError;
-use crate::storage::storage_file::StorageFile;
-use crate::storage::storage_index::StorageIndex;
-use crate::storage::OldStorage;
+use crate::old_storage::storage_file::StorageFile;
+use crate::old_storage::storage_index::StorageIndex;
+use crate::old_storage::OldStorage;
 use std::cell::RefCell;
 use std::rc::Rc;
