+++ src/agdb/lib.rs
@@ -3,8 +3,8 @@ mod commands;
 mod db;
 mod graph;
 mod graph_search;
+mod old_storage;
 mod query;
-mod storage;
 mod transaction;
 mod utilities;
