+++ src/agdb/old_storage.rs
@@ -10,7 +10,7 @@ mod storage_records;
 mod write_ahead_log;
 
 use crate::db::db_error::DbError;
-use crate::storage::storage_index::StorageIndex;
+use crate::old_storage::storage_index::StorageIndex;
 use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::serialize::Serialize;
 use crate::DbIndex;
